requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv
gunicorn